A new version of tdisk.f was added, per revisions made by H.H. Kieffer.
-- description copied from a 2006 May 08  e-mail from H.H. Kieffer

I have pushed a new version of tdisk.f into my area on east...

It is a clean-up of the system of allocating space for the type 5x
files, and a precursor in a concept of begin able to access the bin5
array from a calling program, which would ultimately enable treating
KRC as a non-linear fitting routine! This version may be functionally
little different from what you are now using.

For type 52 and 54, the number of cases that can be stored depends on
how many latitudes (and seasons for Type 52) used. In the print file,
at the end of the first case, will be a line like

KOMMON,KASE,MASE,MTOT= 583200 7400 78 577200

The third item is the number of cases that can be stored at one time,
This example was for type 54 with one latitude, which now allows up to
78 cases.

Type 54 now contains 5 items:  Tsur 1am, Tsur 1pm, heat flow, frost, Tbottom

--end of e-mail



git-svn-id: svn+ssh://bastion.mars.asu.edu/mars/tes/svn/dev/krc/trunk@6581 dedd6328-c50f-0410-8be5-b54a49638459